+++ src/Routes.tsx
@@ -5,7 +5,8 @@ import LandingPage from './pages/landing/LandingPage';
 import LoginPage from './pages/login/LoginPage';
 import PreviousPage from './pages/archive/PreviousPage';
 import ArchiveDetailPage from './pages/archive/ArchiveDetailPage';
-// import SettingsPage from './pages/setting/SettingsPage';
+import { SettingsPage } from './pages/setting/SettingsPage';
+//import SettingsPage from './pages/setting/OldSettingsPage';
 import { SignUp1Page } from './pages/signUp/SignUp1Page';
 //import SignUp1Page from './pages/signUp/old/OldSignUp1Page';
 import SignUp2Page from './pages/signUp/SignUp2Page';
@@ -32,7 +33,7 @@ export const authRoutes = [
   { path: '/myboard', element: <MyBoardPage /> },
   { path: '/archive/:majorName', element: <ArchiveDetailPage /> },
   // 상태관리 도입 예정으로, 잠시 삭제
-  // { path: '/settings', element: <SettingsPage selected={selected} setSelected={setSelected} /> },
+  { path: '/settings', element: <SettingsPage/> },
   { path: '/delete', element: <DeletePage /> },
 ];
